r9260: only overwrite destination files when the temp files was created correct
metze
(This used to be commit b5626c978fe3349f0972e1aeea51b05e66c6ce33)

+++ source4/script/lex_compile.sh
@@ -18,7 +18,9 @@ if [ -r $DEST ]; then
 fi
 TOP=`pwd`
 if cd $dir && $LEX $file; then
-	sed '/^#/ s|$base.yy\.c|$DEST|' $base.yy.c > $base.c
-	rm -f $base.yy.c
+	if [ -r $base.yy.c ];then
+		sed '/^#/ s|$base.yy\.c|$DEST|' $base.yy.c > $base.c
+		rm -f $base.yy.c
+	fi
 fi
 cd $TOP
